requirements.txt
numpy
pandas
scikit-learn
pyyaml
datasets
transformers
triton
trl
peft
torch
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
xformers
